refactored the folder str
