+++ backend/schemas/blog.py
@@ -1,0 +1,26 @@
+from pydantic import BaseModel
+
+
+class BlogBase(BaseModel):
+    title: str
+    sub_title: str
+    content: str
+    image: str
+    user_id: int
+
+
+class BlogCreate(BlogBase):
+    pass
+
+
+class BlogUpdate(BlogBase):
+    id: int
+
+
+class Blog(BlogBase):
+    id: int
+    clap_count: int
+    comment_count: int
+
+    class Config:
+        orm_model = True
